Create prime_factor.cpp
Most of the code is very close to the analogous C code with a couple obvious major advantages: strings exists, and error handling. I don't actually have to do most of that so it works for me. C++ is probably my new favourite language and I look forward to doing more with it in the future.